delete old breadcrumbs
