--- stanzas/breadcrumbs-lit/templates/stanza.html.hbs
@@ -1,1 +1,0 @@
-<div id="breadcrumbs"></div>
